refactor(stella-cli): move reload_from_disk to config/reload.rs; retighten the file-size baseline
reload_from_disk pushed config.rs (1498 on main) over the 1500 ceiling,
and the baseline accepts no new entries — the mutation gets its own
submodule instead. Regenerating the baseline also drops the stale
stella-protocol/src/event.rs entry (its tests moved to event/tests.rs on
main, leaving the file at 1454) and retightens command_deck.rs and the
pipeline entries; AGENTS.md's god-file table and the stella-protocol
README follow the baseline, as check-god-files requires.
